# Speed up input weights visualization.