Speedup input weights visualization by precalculating masks for the first input neuron first.

Results are impressive. Simple bash timing with `time python visualize.py my_net.nnue --features="HalfKP" --save-dir=. --no-hist --no-biases --no-fc-weights --dont-show` reveals following wall-clock times.

num input neurons | master | this PR
---- | ---- | ----
256 | 28 s | 2.8 s
512 | 53 s | 3.0 s